# [Issue 15271] - Get product variant by id