--- a/src/Sylius/Bundle/ProductBundle/Form/EventSubscriber/ProductOptionFieldSubscriber.php
+++ b/src/Sylius/Bundle/ProductBundle/Form/EventSubscriber/ProductOptionFieldSubscriber.php
@@ -44,7 +44,7 @@ final class ProductOptionFieldSubscriber implements EventSubscriberInterface
         $form = $event->getForm();
 
         /** Options should be disabled for configurable product if it has at least one defined variant */
-        $disableOptions = (null !== $this->variantResolver->getVariant($product)) && $product->hasVariants();
+        $disableOptions = null !== $this->variantResolver->getVariant($product);
 
         $form->add('options', ProductOptionAutocompleteType::class, [
             'required' => false,