--- a/src/Sylius/Component/Product/Resolver/DefaultProductVariantResolver.php
+++ b/src/Sylius/Component/Product/Resolver/DefaultProductVariantResolver.php
@@ -19,12 +19,23 @@ use Sylius\Component\Product\Model\ProductVariantInterface;
 
 final class DefaultProductVariantResolver implements ProductVariantResolverInterface
 {
-    public function __construct(private ProductVariantRepositoryInterface $productVariantRepository)
+    public function __construct(private ?ProductVariantRepositoryInterface $productVariantRepository = null)
     {
+        if(!$this->productVariantRepository) {
+            trigger_deprecation('sylius/product', '1.13', 'Not passing a service that implements "%s" as a 1st argument of "%s" constructor is deprecated and will be prohibited in 2.0.', ProductVariantRepositoryInterface::class, self::class);
+        }
     }
 
     public function getVariant(ProductInterface $subject): ?ProductVariantInterface
     {
+        if (!$this->productVariantRepository) {
+            if ($subject->getEnabledVariants()->isEmpty()) {
+                return null;
+            }
+
+            return $subject->getEnabledVariants()->first();
+        }
+
         /** @var ProductVariantInterface|null $productVariant */
         $productVariant = $this->productVariantRepository->findOneBy([
             'product' => $subject->getId(),