--- a/src/Sylius/Component/Product/Resolver/DefaultProductVariantResolver.php
+++ b/src/Sylius/Component/Product/Resolver/DefaultProductVariantResolver.php
@@ -21,27 +21,27 @@ final class DefaultProductVariantResolver implements ProductVariantResolverInter
 {
     public function __construct(private ?ProductVariantRepositoryInterface $productVariantRepository = null)
     {
-        if(!$this->productVariantRepository) {
+        if (null === $this->productVariantRepository) {
             trigger_deprecation('sylius/product', '1.13', 'Not passing a service that implements "%s" as a 1st argument of "%s" constructor is deprecated and will be prohibited in 2.0.', ProductVariantRepositoryInterface::class, self::class);
         }
     }
 
     public function getVariant(ProductInterface $subject): ?ProductVariantInterface
     {
-        if (!$this->productVariantRepository) {
-            if ($subject->getEnabledVariants()->isEmpty()) {
-                return null;
-            }
+        if ($this->productVariantRepository && $subject->getId()) {
+            /** @var ProductVariantInterface|null $productVariant */
+            $productVariant = $this->productVariantRepository->findOneBy([
+                'product' => $subject->getId(),
+                'enabled' => true,
+            ]);
 
-            return $subject->getEnabledVariants()->first();
+            return $productVariant;
         }
 
-        /** @var ProductVariantInterface|null $productVariant */
-        $productVariant = $this->productVariantRepository->findOneBy([
-            'product' => $subject->getId(),
-            'enabled' => true,
-        ]);
+        if ($subject->getEnabledVariants()->isEmpty()) {
+            return null;
+        }
 
-        return $productVariant;
+        return $subject->getEnabledVariants()->first();
     }
 }