--- a/src/Sylius/Component/Product/Resolver/DefaultProductVariantResolver.php
+++ b/src/Sylius/Component/Product/Resolver/DefaultProductVariantResolver.php
@@ -13,17 +13,24 @@ declare(strict_types=1);
 
 namespace Sylius\Component\Product\Resolver;
 
+use Sylius\Component\Product\Repository\ProductVariantRepositoryInterface;
 use Sylius\Component\Product\Model\ProductInterface;
 use Sylius\Component\Product\Model\ProductVariantInterface;
 
 final class DefaultProductVariantResolver implements ProductVariantResolverInterface
 {
+    public function __construct(private ProductVariantRepositoryInterface $productVariantRepository)
+    {
+    }
+
     public function getVariant(ProductInterface $subject): ?ProductVariantInterface
     {
-        if ($subject->getEnabledVariants()->isEmpty()) {
-            return null;
-        }
+        /** @var ProductVariantInterface|null $productVariant */
+        $productVariant = $this->productVariantRepository->findOneBy([
+            'product' => $subject->getId(),
+            'enabled' => true,
+        ]);
 
-        return $subject->getEnabledVariants()->first();
+        return $productVariant;
     }
 }