--- a/src/Sylius/Component/Product/spec/Resolver/DefaultProductVariantResolverSpec.php
+++ b/src/Sylius/Component/Product/spec/Resolver/DefaultProductVariantResolverSpec.php
@@ -22,22 +22,38 @@ use Sylius\Component\Product\Resolver\ProductVariantResolverInterface;
 
 final class DefaultProductVariantResolverSpec extends ObjectBehavior
 {
-    function let(ProductVariantRepositoryInterface $productVariantRepository): void
-    {
-        $this->beConstructedWith($productVariantRepository);
-    }
-
     function it_implements_variant_resolver_interface(): void
     {
         $this->shouldImplement(ProductVariantResolverInterface::class);
     }
 
-    function it_returns_first_variant(
+    function it_returns_first_variant_if_product_variant_repository_is_not_initialized(
         ProductInterface $product,
         ProductVariantInterface $variant,
         Collection $variants,
+    ): void {
+        $product->getEnabledVariants()->willReturn($variants);
+        $variants->isEmpty()->willReturn(false);
+        $variants->first()->willReturn($variant);
+
+        $this->getVariant($product)->shouldReturn($variant);
+    }
+
+    function it_returns_null_if_first_variant_is_not_defined_and_product_variant_repository_is_not_initialized(Collection $variants, ProductInterface $product): void
+    {
+        $product->getEnabledVariants()->willReturn($variants);
+        $variants->isEmpty()->willReturn(true);
+
+        $this->getVariant($product)->shouldReturn(null);
+    }
+
+    function it_returns_first_variant_if_product_variant_repository_is_initialized(
+        ProductInterface $product,
+        ProductVariantInterface $variant,
         ProductVariantRepositoryInterface $productVariantRepository,
     ): void {
+        $this->beConstructedWith($productVariantRepository);
+
         $product->getId()->willReturn(1);
         $productVariantRepository->findOneBy([
             'product' => 1,
@@ -47,12 +63,13 @@ final class DefaultProductVariantResolverSpec extends ObjectBehavior
         $this->getVariant($product)->shouldReturn($variant);
     }
 
-    function it_returns_null_if_first_variant_is_not_defined(
-        Collection $variants,
+    function it_returns_null_if_first_variant_is_not_defined_and_product_variant_repository_is_initialized(
         ProductInterface $product,
         ProductVariantRepositoryInterface $productVariantRepository,
     ): void
     {
+        $this->beConstructedWith($productVariantRepository);
+
         $product->getId()->willReturn(1);
         $productVariantRepository->findOneBy([
             'product' => 1,