@jakubtobiasz 2023-09-05
Hello @ulinskas 👋🏼,
do you have any statistics comparing the performance before and after your changes?🙏🏼