@jakubtobiasz 2023-09-05
We're aware it'll be noticeable in case of bigger amount of data. It'd be great if you share some numbers 👊🏼.