@ulinskas 2023-09-06
@jakubtobiasz 
Here are some simple numbers of variants and performance changes.
I used symfony profiler for simple performance benchmarks.

This is before and after my implementation.
Product with 12 variants:
   controller execution 365.1 ms / 20 MiB -> **controller execution 397 ms / 18 MiB**
Product with 1229 variants:
   controller execution 450.9 ms / 30 MiB -> **controller execution 338.3 ms / 24 MiB**
Product with 38775 variants:
    controller execution 2006.2 ms / 222 MiB -> **controller execution 372.7 ms / 14 MiB**
Product with 97225 variants:
   Memory limit reached -> **controller execution 358.6 ms / 16 MiB**
      
As you can see, most noticeable performance gains are for products with 10k+ variants.
For products with few variants, there aren't any noticeable changes, for each request execution time differe by +-40ms.
Keep in mind, I'm using local docker environment for mac, so this is not from production.

Also attaching screenshots, for total page load performance. Screenshots are from product with 38775 variants.
Before:
<img width="1054" alt="38775-before" src="https://github.com/Sylius/Sylius/assets/17234071/507f8105-d2fb-438b-a441-dded6d22492e">
After:
<img width="1022" alt="38775-after" src="https://github.com/Sylius/Sylius/assets/17234071/9de8ec18-4b75-4f25-9b66-43226b704db0">

Hope this helps.