@diimpp 2023-09-13
Basically every other relation with collection will go out of memory on 1000+ records, by adding optimization at `Component` level we're making assumptions about `Bundle` mappings and `Core`/`Admin` pages/usage.

For example `Order->getLastPayment(?string $state = null): ?PaymentInterface` or `Product->getImagesByType(string $type): Collection` will go out of memory on 1000-9000 records as well.

I would say we should move this change into `Sylius\Bundle\CoreBundle\Product\Resolver\ProductVariantResolver`as decorating service. (Or `Component\Core`, not sure about the difference atm) to keep optimization to "ecommerce-application" level instead of "ecommerce-framework" level, where it's actually needed based on usage.