@diimpp 2023-09-29
@GSadee shouldn't this go into 1.13 instead of 1.12 as this is not a regular bugfix? Then deprecation notice can be added immediately. 

Similar performance fix for zone matcher went into 1.13 https://github.com/Sylius/Sylius/pull/15275