@GSadee 2023-09-29
> @GSadee shouldn't this go into 1.13 instead of 1.12 as this is not a regular bugfix? Then deprecation notice can be added immediately.
> 
> Similar performance fix for zone matcher went into 1.13 #15275

I have been thinking about this and decided that we could treat it as a bug because of reaching the memory limit with a large number of variants, fix it in this version and add the appropriate deprecation on 1.13. Are you having doubts that we should not consider this a bug?