@diimpp 2023-09-29
> Are you having doubts that we should not consider this a bug?

No, personally I don't consider this a bug, every collection relation is susceptible to this issue, try adding 1000 items to a order, 2000 payment attempts for a order, it doesn't make order system broken because of that.

So in my opinion it's not a bugfix for a degraded performance, but rather a feature/improvement to support "bigger" data sets.